requirements.txt
requests>=2.31.0
pyyaml>=6.0
pytest>=7.4.0
